TugasBesar2/requirements.txt
pycryptodome
cryptography
